Fix operator bug
CLicking more operators after the first would affect the sum output.
The equals logic was only recognising the sum as an array of length 1 when this bug occured, due to an incorrect updating of operatorVal and operatorSymbol in the operator eventListener.
This led to the default logic occurring when there is only 1 recognised element in sum array, which is to continue to display the same number. In this case, it looked like the sum was still present, as the original operator was now part of the string num1.